simplify env fns
