+++ lib/modules/datasource/crate/index.ts
@@ -4,7 +4,7 @@ import { GlobalConfig } from '../../../config/global';
 import { logger } from '../../../logger';
 import * as memCache from '../../../util/cache/memory';
 import { cache } from '../../../util/cache/package/decorator';
-import { getChildEnv } from '../../../util/exec/utils';
+import { getEnv } from '../../../util/env';
 import { privateCacheDir, readCacheFile } from '../../../util/fs';
 import { simpleGitConfig } from '../../../util/git/config';
 import { toSha256 } from '../../../util/hash';
@@ -323,7 +323,7 @@ export class CrateDatasource extends Datasource {
         const git = Git({
           ...simpleGitConfig(),
           maxConcurrentProcesses: 1,
-        }).env(getChildEnv());
+        }).env(getEnv());
         const clonePromise = git.clone(registryFetchUrl, clonePath, {
           '--depth': 1,
         });
