+++ lib/modules/datasource/git-refs/base.ts
@@ -2,7 +2,7 @@ import is from '@sindresorhus/is';
 import { simpleGit } from 'simple-git';
 import { logger } from '../../../logger';
 import { cache } from '../../../util/cache/package/decorator';
-import { getChildEnv } from '../../../util/exec/utils';
+import { getEnv } from '../../../util/env';
 import { getGitEnvironmentVariables } from '../../../util/git/auth';
 import { simpleGitConfig } from '../../../util/git/config';
 import { getRemoteUrlWithToken } from '../../../util/git/url';
@@ -34,7 +34,7 @@ export abstract class GitDatasource extends Datasource {
     const gitSubmoduleAuthEnvironmentVariables = getGitEnvironmentVariables([
       this.id,
     ]);
-    const gitEnv = getChildEnv({ env: gitSubmoduleAuthEnvironmentVariables });
+    const gitEnv = getEnv(false, { env: gitSubmoduleAuthEnvironmentVariables });
     const git = simpleGit(simpleGitConfig()).env(gitEnv);
 
     // fetch remote tags
